Moved mock classes from the root mock package into test packages
