+++ observation/observationtest/error_reader.go
@@ -1,4 +1,4 @@
-package mock
+package observationtest
 
 import "io"
